[project @ Compile fix, cast to int]
Original author: Philipp Kern <phil@0x539.de>
Date: 2005-03-30 16:36:57+00:00

+++ src/line.cpp
@@ -250,7 +250,8 @@ net6::packet obby::line::to_packet(unsigned int document_id) const
 
 	std::vector<user_pos>::const_iterator iter;
 	for(iter = m_authors.begin(); iter != m_authors.end(); ++ iter)
-		pack << iter->position << iter->author->get_id();
+		pack << static_cast<int>(iter->position) <<
+			static_cast<int>(iter->author->get_id());
 
 	return pack;
 }
